Split the menu screen into a reloadable render boundary
Reshape `src/screens/menu` into a hot-reload-friendly layout: a pure
`render` path (the reload target) that reads everything from a
host-prepared `HostContext`, separated `state`, and a `render_hot_stub`
the in-process build links when the live cdylib is absent. Supporting
tweaks to `visual_style_bg` and the menu bench keep the boundary pure.
